+++ Examples/Examples-iOS/Pods/Target Support Files/IGListKit/IGListKit-umbrella.h
@@ -31,7 +31,8 @@
 #import "IGListSupplementaryViewSource.h"
 #import "IGListUpdatingDelegate.h"
 #import "IGListWorkingRangeDelegate.h"
-#import "NSObject+IGListDiffable.h"
+#import "NSNumber+IGListDiffable.h"
+#import "NSString+IGListDiffable.h"
 
 FOUNDATION_EXPORT double IGListKitVersionNumber;
 FOUNDATION_EXPORT const unsigned char IGListKitVersionString[];